Now make_ping_time() in common/util.c returns the time in milliseconds instead of microseconds. In fe_set_lag(serv, lag) the lag argument is now in milliseconds as well.
Fixes #758.

+++ src/common/fe.h
@@ -118,7 +118,7 @@ void fe_set_nick (struct server *serv, char *newnick);
 void fe_ignore_update (int level);
 void fe_beep (session *sess);
 void fe_lastlog (session *sess, session *lastlog_sess, char *sstr, gtk_xtext_search_flags flags);
-void fe_set_lag (server *serv, int lag);
+void fe_set_lag (server *serv, long lag);
 void fe_set_throttle (server *serv);
 void fe_set_away (server *serv);
 void fe_serverlist_open (session *sess);
